Several clean up around creating new isa fundraise, incorporating it into end to end test

+++ app/data/deleteFundraiseData.server.ts
@@ -1,5 +1,9 @@
 import getMysql from "./mysql.server";
-import { NotFoundError, MethodNotAllowedError } from "aws-sdk-plus/dist/errors";
+import {
+  NotFoundError,
+  MethodNotAllowedError,
+  ConflictError,
+} from "aws-sdk-plus/dist/errors";
 
 const deleteFundraiseData = ({
   uuid,
@@ -9,13 +13,23 @@ const deleteFundraiseData = ({
   userId: string;
 }) =>
   getMysql().then(({ execute, destroy }) => {
-    return execute(`SELECT c.userId FROM contract c WHERE c.uuid = ?`, [uuid])
-      .then((results) => {
-        const [contract] = results as { userId: string }[];
+    return execute(`SELECT c.userId, c.uuid FROM contract c WHERE c.uuid = ?`, [
+      uuid,
+    ])
+      .then(async (results) => {
+        const [contract] = results as { userId: string; uuid: string }[];
         if (!contract)
           throw new NotFoundError(`Could not find contract ${uuid}`);
         if (userId !== contract.userId)
           throw new MethodNotAllowedError(`Not authorized to delete ${uuid}`);
+        const eversignDocs = await execute(
+          `SELECT e.id FROM eversigndocument e INNER JOIN agreement a ON a.uuid = e.agreementUuid WHERE a.contractUuid = ?`,
+          [contract.uuid]
+        ).then((a) => a as { id: string }[]);
+        if (eversignDocs.length)
+          throw new ConflictError(
+            `Cannot delete a fundraise with live legal documents.`
+          );
         return Promise.all([
           execute(`DELETE FROM contractdetail WHERE contractUuid = ?`, [uuid]),
           execute(`DELETE FROM agreement WHERE contractUuid = ?`, [uuid]),
